feat: create membership.md [DONT MERGE] [NEEDS DISCUSSION] (#317)
* feat: create membership.md

Initial cut of documented governance, this is long overdue.

* fix: remove table, it's just confusing for now

* fix: remove more evidence that I just copied and pasted this from k8s

* feat: Codify the Blaede. 

We've now successfully wielded the Blaede twice. It came with a lore card, it reads as follows... 

The Blaede itself is inscribed with: 

May the pattern succeed us all.yaml

* fix: specify ownership as a github structure

* feat: add needs of the project

So that if there's a neglected area people can be encouraged to take it over and do awesome things without waiting.