fix: SELinux to permissive and added stopping the...
Waydroid container after the Google Play Certificate has been added.  It's better practice to have SELinux in permissive mode rather than disabled.  I know this is an upstream bug with SELinux and OCI, but for now it's better to have users just have it set to permissive then to fully disable it.