chore: SELinux in permissive.
I have tested that this works fine in permissive as opposed to disabling it.  Disabling it is not recommended, and I am aware that permissive only logs any issues it has, it's a lot better than disabling it.  This is especially the case if the user plans on re-enabling it after the SELinux issue has been resolved upstream.