feat: Add "Wallpapers" page for website
I originally wanted to create a custom `just` command for Bazzite to bring back Universal Blue wallpapers, since they were removed to save on storage which is understandable.

I decided this was easier for me to accomplish, but is probably not necessary.

I also think the repository should have directories dedicated to Bluefin, Bazzite, and Beyond eventually.  However maybe not until wallpapers are made for them.