feat: Steam hybrid graphics fix & some corrections (#559)
* feat: Steam hybrid graphics fix & some corrections

I understand this is out of scope, but it has been reported enough times where I feel it needs to be addressed, so it can be linked.  I will remove this as soon as Valve fixes their desktop files for the Linux Steam client.

* chore: more information on why use Fedora?

* chore: update the path to be universal

* feat: Add Steam Deck update indicator answer

I have seen a lot of questions surrounding this.