chore: use `--pull=newer` for create-ublue-image
As the create-ublue-image tool evolves, it is possible that users
could end up running an older version of the tool when setting up
a new custom ublue image. Supplying `--pull=newer` to the
`podman run` command ensures that users should always get the
latest image when the command is run.

Signed-off-by: Micah Abbott <miabbott@redhat.com>